form-mapper/requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==8.2.0
pytest-cov==4.1.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
pytest-playwright==0.4.8

# Linting
flake8==7.1.1
flake8-docstrings==1.8.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0

# Formatting
black==24.4.2
isort==5.13.2

# Type checking
mypy==1.10.0
types-python-dateutil==2.8.19.14
typing-extensions>=4.0.0

# Documentation
sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==2.1.0

# Development tools
pre-commit==3.7.0
bump2version==1.0.1
twine==5.0.0
